+++ src/request.rs
@@ -4,6 +4,7 @@
 use serde::Deserialize;
 use serde::Serialize;
 
+use crate::args::RemindAt;
 use crate::args::RemindIn;
 use crate::channel::Writer;
 use crate::reminder::Reminder;
@@ -13,6 +14,12 @@ use crate::reminder::Reminder;
 pub(crate) enum Request {
   Remind(Reminder),
   List(Writer<Vec<Reminder>>),
+}
+
+impl From<RemindAt> for Request {
+  fn from(other: RemindAt) -> Self {
+    Self::Remind(Reminder::from(other))
+  }
 }
 
 impl From<RemindIn> for Request {